requirements.txt
django==4.2.11
django-environ==0.11.2
djangorestframework==3.15.0
drf_spectacular==0.27.1
psycopg[c]==3.1.18
wagtailstreamforms==4.1.0
wagtail-localize==1.8.2
wagtail==5.2.3
wagtail-modeladmin==2.0.0